+++ app/services/index.ts
@@ -8,7 +8,7 @@ export * from "./app-insights";
 export * from "./autoscale-formula.service";
 export * from "./azure-http.service";
 export * from "./arm-http.service";
-export * from "./batch-labs.service";
+export * from "./batch-explorer.service";
 export * from "./cache-data.service";
 export * from "./compute.service";
 export * from "./github-data";
