+++ app/services/themes/theme.service.ts
@@ -8,7 +8,7 @@ const stripJsonComments = require("strip-json-comments");
 import { FileSystemService, NotificationService } from "@batch-flask/ui";
 import { log } from "app/utils";
 import { filter } from "rxjs/operators";
-import { BatchExplorerService } from "../batch-labs.service";
+import { BatchExplorerService } from "app/services/batch-explorer.service";
 import { SettingsService } from "../settings.service";
 import { Theme } from "./theme.model";
